requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
langdetect>=1.0.9
youtube-transcript-api>=0.6.2
biopython>=1.81
